Add search options to holdings segment.
